code review, try movign timing to script
